refactor: move data models to schemas package
